requirements.txt
# Dependencias principales de IA y LLM
openai>=1.10.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-text-splitters>=0.1.0
langchain-community>=0.0.10
langchain-tavily>=0.1.0

# Interfaz de usuario
gradio>=4.10.0

# Procesamiento de datos
pandas>=2.0.0
numpy>=1.24.0

# Generación de PDFs
fpdf2>=2.7.6
reportlab>=4.0.0

# Procesamiento de documentos
PyPDF2>=3.0.0
python-docx>=0.8.11

# Web scraping y requests
requests>=2.31.0
beautifulsoup4>=4.12.0

# Análisis de texto y NLP
spacy>=3.7.0
sentence-transformers>=2.2.2

# Visualización y gráficos
matplotlib>=3.7.3
seaborn>=0.13.0
Pillow>=10.0.0

# Utilidades científicas
scikit-learn>=1.3.0

# Utilidades de desarrollo
python-dotenv>=1.0.0
tqdm>=4.66.0

# Variables de entorno y configuración
pydantic>=2.0.0

# Para logging avanzado
colorlog>=6.7.0 